refactor step 9 - cleanup before merge
