+++ app/src/main/java/com/replica/replicaisland/rendering/RenderSystem.kt
@@ -1,5 +1,6 @@
 package com.replica.replicaisland.rendering
 
+import com.replica.replicaisland.GameRenderer
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.rendering.DrawableObject
 import com.replica.replicaisland.utils.FixedSizeArray
